[simulation][gaming][protocol][integration] dev for SWBUG-11120 "Scipio - sur mission recueillir, la ligne de receueil ne remonte pas l'info des unités la traversant"
- merges from Scipio : pull requests concerning SWBUG-11120 and SWBUG-8259 "Scipio 380420 - PSG4_447 Mission MELEE ARMER UN PIA" (https://github.com/masagroup/sw.sword/pull/393, https://github.com/masagroup/sw.sword/pull/406, https://github.com/masagroup/sw.sword/pull/449 ): new action "PHY_ActionRecoLima"
- refactoring lima function enum (E_FuncLimaType was duplicated) + new report parameter to allow phase line type to be passed.

+++ src/libraries/simulation_kernel/Entities/Orders/MIL_Report.h
@@ -279,7 +279,7 @@ template< typename T >
 void MIL_Report::PostEvent( const T& receiver, const MIL_DecisionalReport& nReport, const MIL_LimaFunction& limaFunction, const std::string& nParam )
 {
     std::vector< boost::shared_ptr< MIL_MissionParameter_ABC > > parameters;
-    auto pParameter1 = MIL_MissionParameterFactory::CreateLimaFunction( static_cast< int >( limaFunction.GetID() ) );
+    auto pParameter1 = MIL_MissionParameterFactory::CreateLimaFunction( limaFunction.GetID() );
     auto pParameter2 = MIL_MissionParameterFactory::Create( nParam );
     parameters.push_back( pParameter1 );
     parameters.push_back( pParameter2 );
